Refactored Block Manger iterator and Block postions
New Iterator objects now takes the 2d array of blocks and a length parameter and iterates over the whole array.
It's now up to the board to work out what to do with the block and whether it is visible.

Block no longer knows where it is in the array. It's now up to the Block manager and the iterator to report this.